requirements-api.txt
fastapi
uvicorn
pandas
psycopg2-binary
requests
python-multipart